refactor: 통계 화면 ViewModel을 StatsViewModel로 통합
